🚚 api: move controller method to src
