+++ api/lib/application/certification-candidates/certification-candidates-controller.js
@@ -1,17 +1,5 @@
 import { usecases } from '../../domain/usecases/index.js';
 
-const authorizeToStart = async function (request, h) {
-  const certificationCandidateForSupervisingId = request.params.id;
-
-  const authorizedToStart = request.payload['authorized-to-start'];
-  await usecases.authorizeCertificationCandidateToStart({
-    certificationCandidateForSupervisingId,
-    authorizedToStart,
-  });
-
-  return h.response().code(204);
-};
-
 const authorizeToResume = async function (request, h) {
   const certificationCandidateId = request.params.id;
 
@@ -31,7 +19,6 @@ const endAssessmentBySupervisor = async function (request) {
 };
 
 const certificationCandidatesController = {
-  authorizeToStart,
   authorizeToResume,
   endAssessmentBySupervisor,
 };
