+++ api/tests/certification/enrolment/unit/application/certification-candidate-controller_test.js
@@ -1,3 +1,4 @@
+import { usecases as libUsecases } from '../../../../../lib/domain/usecases/index.js';
 import { certificationCandidateController } from '../../../../../src/certification/enrolment/application/certification-candidate-controller.js';
 import { usecases } from '../../../../../src/certification/enrolment/domain/usecases/index.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
@@ -45,6 +46,37 @@ describe('Unit | Controller | certification-candidate-controller', function () {
         },
       });
       expect(response.statusCode).to.equal(201);
+    });
+  });
+
+  describe('#authorizeToStart', function () {
+    it('should return a 204 status code', async function () {
+      // given
+      const request = {
+        auth: {
+          credentials: { userId: '111' },
+        },
+        params: {
+          id: 99,
+        },
+        payload: { 'authorized-to-start': true },
+      };
+
+      sinon.stub(libUsecases, 'authorizeCertificationCandidateToStart');
+
+      libUsecases.authorizeCertificationCandidateToStart = sinon.stub().rejects();
+      libUsecases.authorizeCertificationCandidateToStart
+        .withArgs({
+          certificationCandidateForSupervisingId: 99,
+          authorizedToStart: true,
+        })
+        .resolves();
+
+      // when
+      const response = await certificationCandidateController.authorizeToStart(request, hFake);
+
+      // then
+      expect(response.statusCode).to.equal(204);
     });
   });
