+++ api/tests/unit/application/certification-candidates/certification-candidates-controller_test.js
@@ -3,35 +3,6 @@ import { usecases } from '../../../../lib/domain/usecases/index.js';
 import { expect, hFake, sinon } from '../../../test-helper.js';
 
 describe('Unit | Controller | certifications-candidate-controller', function () {
-  describe('#authorizeToStart', function () {
-    it('should return a 204 status code', async function () {
-      // given
-      const request = {
-        auth: {
-          credentials: { userId: '111' },
-        },
-        params: {
-          id: 99,
-        },
-        payload: { 'authorized-to-start': true },
-      };
-
-      usecases.authorizeCertificationCandidateToStart = sinon.stub().rejects();
-      usecases.authorizeCertificationCandidateToStart
-        .withArgs({
-          certificationCandidateForSupervisingId: 99,
-          authorizedToStart: true,
-        })
-        .resolves();
-
-      // when
-      const response = await certificationCandidatesController.authorizeToStart(request, hFake);
-
-      // then
-      expect(response.statusCode).to.equal(204);
-    });
-  });
-
   describe('#authorizeToResume', function () {
     it('should return a 204 status code', async function () {
       // given
